+++ rivershared/testsignal/test_signal_test.go
@@ -2,11 +2,9 @@ package testsignal
 
 import (
 	"testing"
-	"time"
 
 	"github.com/stretchr/testify/require"
 
-	"github.com/riverqueue/river/rivershared/riversharedtest"
 	"github.com/riverqueue/river/rivershared/util/testutil"
 )
 
@@ -106,13 +104,4 @@ func TestTestSignal(t *testing.T) {
 
 		signal.WaitOrTimeout()
 	})
-}
-
-// Marked as non-parallel because `t.Setenv` is not compatible with `t.Parallel`.
-func TestWaitTimeout(t *testing.T) {
-	t.Setenv("GITHUB_ACTIONS", "")
-	require.Equal(t, 3*time.Second, riversharedtest.WaitTimeout())
-
-	t.Setenv("GITHUB_ACTIONS", "true")
-	require.Equal(t, 10*time.Second, riversharedtest.WaitTimeout())
 }
